# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NoorPath - Muslim Parenting & Quran Learning App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories/contextual:
    get:
      operationId: listContextualStories
      tags: [stories]
      summary: List stories related to a surah or ayah
      parameters:
        - name: surahNumber
          in: query
          required: true
          schema:
            type: integer
        - name: ayah
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Contextual story suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: object
                    properties:
                      surahNumber:
                        type: integer
                      ayah:
                        type: integer
                        nullable: true
                    required: [surahNumber]
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                required: [context, stories]
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List Islamic stories
      parameters:
        - name: ageGroup
          in: query
          schema:
            type: string
            enum: [toddler, child, preteen, teen]
        - name: storyType
          in: query
          schema:
            type: string
            enum: [quranic_narrative, seerah_context, companion_profile, moral_lesson]
        - name: surahNumber
          in: query
          schema:
            type: integer
        - name: ayah
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                required: [stories]
  /stories/{storyIdOrSlug}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get a specific story
      parameters:
        - name: storyIdOrSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryDetail"
...
components:
  schemas:
...
    StoryType:
      type: string
      enum: [quranic_narrative, seerah_context, companion_profile, moral_lesson]
    StorySources:
      type: object
      properties:
        primary:
          type: string
        hadith:
          type: array
          items:
            type: string
        seerah:
          type: array
          items:
            type: string
        notes:
          type: string
    StoryAyahRef:
      type: object
      properties:
        surahNumber:
          type: integer
        ayahStart:
          type: integer
        ayahEnd:
          type: integer
        label:
          type: string
      required: [surahNumber, ayahStart, ayahEnd]
...
    Story:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        storyType:
          $ref: "#/components/schemas/StoryType"
        ageGroup:
          type: string
          enum: [toddler, child, preteen, teen]
        summary:
          type: string
        readingTimeMinutes:
          type: integer
        featuredCharacter:
          type: string
        morals:
          type: array
          items:
            type: string
        relatedAyahs:
          type: array
          items:
            $ref: "#/components/schemas/StoryAyahRef"
        sources:
          $ref: "#/components/schemas/StorySources"
      required: [id, slug, title, storyType, ageGroup, summary, readingTimeMinutes, featuredCharacter, morals, relatedAyahs, sources]
    StoryDetail:
      allOf:
        - $ref: "#/components/schemas/Story"
        - type: object
          properties:
            content:
              type: string
            discussionQuestions:
              type: array
              items:
                type: string
          required: [content, discussionQuestions]
...