# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoorPath - Muslim Parenting & Quran Learning App API
 * OpenAPI spec version: 0.1.0
 */
import type { StoryAgeGroup } from "./storyAgeGroup";
import type { StoryAyahRef } from "./storyAyahRef";
import type { StorySources } from "./storySources";
import type { StoryType } from "./storyType";

export interface Story {
  id: number;
  slug: string;
  title: string;
  storyType: StoryType;
  ageGroup: StoryAgeGroup;
  summary: string;
  readingTimeMinutes: number;
  featuredCharacter: string;
  morals: string[];
  relatedAyahs: StoryAyahRef[];
  sources: StorySources;
}