# lib/api-zod/src/generated/types/storySources.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoorPath - Muslim Parenting & Quran Learning App API
 * OpenAPI spec version: 0.1.0
 */

export interface StorySources {
  primary?: string;
  hadith?: string[];
  seerah?: string[];
  notes?: string;
}