# lib/api-zod/src/generated/types/storyType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoorPath - Muslim Parenting & Quran Learning App API
 * OpenAPI spec version: 0.1.0
 */

export type StoryType = (typeof StoryType)[keyof typeof StoryType];

export const StoryType = {
  quranic_narrative: "quranic_narrative",
  seerah_context: "seerah_context",
  companion_profile: "companion_profile",
  moral_lesson: "moral_lesson",
} as const;